+++ cint/cint/src/bc_parse.cxx
@@ -1854,7 +1854,7 @@ int G__blockscope::initstruct(G__TypeReader& type, struct G__var_array* var, int
   // Read and process the initializer specification.
   //
   int mparen = 1;
-  size_t linear_index = -1;
+  long linear_index = -1;
   buf.obj.i = var->p[varid] + memvar->p[memindex];
   G__FastAllocString expr(G__ONELINE);
   while (mparen) {
@@ -1865,7 +1865,7 @@ int G__blockscope::initstruct(G__TypeReader& type, struct G__var_array* var, int
       // FIXME: Do we handle a string literal correctly here?
       ++linear_index;
       // If we are an array, make sure we have not gone beyond the end.
-      if ((num_of_elements || isauto) && (linear_index >= num_of_elements)) {
+      if ((num_of_elements || isauto) && (linear_index >= (long)num_of_elements)) {
         // -- We have gone past the end of the array.
 	if (isauto) {
           // -- Unspecified length array, make it bigger to fit.
@@ -1874,7 +1874,7 @@ int G__blockscope::initstruct(G__TypeReader& type, struct G__var_array* var, int
 	}
 	else {
           // -- Fixed-size array, error, array index out of range.
-	  G__fprinterr(G__serr, "Error: %s: %d: Array initialization out of range *(%s+%d), upto %d ", __FILE__, __LINE__, type.Name(), linear_index, num_of_elements);
+	  G__fprinterr(G__serr, "Error: %s: %d: Array initialization out of range *(%s+%ld), upto %lu ", __FILE__, __LINE__, type.Name(), linear_index, num_of_elements);
 	  G__genericerror(0);
           while (mparen-- && (c1 != ';')) {
             c1 = G__fignorestream("};");
